+++ src/main/java/by/training/shape/comparator/PyramidIDComparator.java
@@ -8,7 +8,7 @@ public class PyramidIDComparator implements Comparator<Pyramid> {
 
     @Override
     public int compare(Pyramid o1, Pyramid o2) {
-        int compare = Double.compare(o1.getId(), o2.getId());
+        int compare = Long.compare(o1.getId(), o2.getId());
         return compare;
     }
 }
