+++ src/main/java/by/training/shape/util/IdGenerator.java
@@ -2,12 +2,12 @@ package by.training.shape.util;
 
 public class IdGenerator {
 
-    private static int id = 0;
+    private static long id = 0L;
 
     private IdGenerator(){
     }
 
-    public static int generateId(){
+    public static long generateId(){
         return ++id;
     }
 }
